backend/requirements.txt
# FastAPI 核心框架 (最新穩定版本)
fastapi==0.116.1
uvicorn[standard]==0.35.0

# 資料庫相關
sqlalchemy==2.0.43
pymysql==1.1.1
alembic==1.16.5

# 認證與安全
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==45.0.6

# 工具庫
python-multipart==0.0.12
pydantic==2.11.7
python-dotenv==1.1.1
websockets==15.0.1

# 設定管理
pydantic-settings==2.7.0